implemented paralax effect and hollow arrow effect
