+++ src/App.css
@@ -1,8 +1,28 @@
 .hero {
-    background-image: url('/hero_background.jpg');
+    background-image: url('/img/hero_background.jpg');
+
+    background-attachment: fixed;
+
     background-position: center;
     background-size: cover;
     background-repeat: no-repeat;
     height: 100vh; /* Full viewport height */
     width: 100%;   /* Full width */
+   
+   
+    position: fixed; /* Keep the content in place */
+    z-index: -10; /* Ensure it's on top of the background */
+    color: white; /* Adjust color for readability */
+    text-align: center;
 }
+
+.hollow-arrow {
+    width: 100%;
+    height: 40px;
+    background-image: url("../img/content_hollow-arrow.png");
+    background-position: 50% 50%;
+    position: relative;
+    top: -40px;
+    left: 0;
+    opacity: 1;
+  }
