+++ src/Pages/Home/HeroSection.tsx
@@ -1,0 +1,15 @@
+
+export default function HeroSection() {
+  return (
+    <section className=" hero hero-content min-h-screen  bg-gradient-to-br from-tpurple to-tblue flex flex-col justify-center items-start text-white">
+    <p className=" text-6xl p-8 px-28">Hi there!</p>
+    <p className="text-4xl p-8 px-28">I am Rino, I bring ideas into digital reality</p>
+    <p className="text-xl p-8 px-28">I am a computer science engineering student with a passion for building software . I believe in continuous learning and am eager to take on new challenges. Take a look at my work and feel free to reach out for a chat.</p>
+    <button className="text-xl mx-auto my-8 p-4 border rounded-md hover:bg-white hover:text-tblue">
+      Contact me
+    </button>
+
+
+  </section>
+  )
+}
