+++ src/Pages/Home/Home.tsx
@@ -1,0 +1,20 @@
+import HeroSection from "./HeroSection";
+
+
+export default function Home() {
+  return (
+    <main className=" min-h-screen ">
+
+<HeroSection/>
+
+<section className=" z-20  min-h-screen w-full ">
+
+</section >
+<section className=" z-20  min-h-screen w-full bg-gradient-to-br from-tpurple to-tblue">
+  <div className="hollow-arrow">
+
+  </div>
+</section>
+    </main>
+  )
+}
